requirements.txt
pandas
scipy  
matplotlib
seaborn
pillow
opencv-contrib-python
tensorflow == 2.5